requirements.txt
fastapi
uvicorn[standard]
websockets
wsproto
pillow
numpy
pydantic
ultralytics
opencv-python
paddlepaddle==3.2.0
paddleocr==3.3.1
python-multipart
python-dotenv
httpx